InformeTorn logic almost completed.

+++ JocsDeGuerra/Models/AccumulatedPoints.cs
@@ -7,9 +7,9 @@ namespace JocsDeGuerra.Models
 {
     public class AccumulatedPoints
     {
-        public int ProductionPoints { get; set; }
-        public int ResearchPoints { get; set; }
-        public int ExplorationPoints { get; set; }
-        public int IndustryPoints { get; set; }
+        public double ProductionPoints { get; set; }
+        public double ResearchPoints { get; set; }
+        public double ExplorationPoints { get; set; }
+        public double IndustryPoints { get; set; }
     }
 }
